Fix all 50 compilation errors in processor crate
This commit resolves all compilation errors in llm-optimizer-processor,
reducing from 50 errors to 0 and making the crate ready for publication.

Key fixes:
- PostgresTransaction: Fixed executor access using &mut **tx.executor()
- Redis PubSub: Refactored to use proper async connection and PubSub API
- AggregationResult: Made constructor and setters public
- Storage traits: Added Send + Sync + 'static bounds to all generic parameters
- Serialization: Added Serialize/DeserializeOwned trait bounds where needed
- Type mismatches: Fixed Duration conversions, usize/u64 casts, Option wrapping
- Method signatures: Fixed RollbackEngine::new(), execute_rollback(), and other call sites
- Trait objects: Changed WatermarkGenerator to concrete types, added Debug impls
- Borrow checker: Cloned values and restructured code to avoid conflicts
- OpenTelemetry: Updated to use correct BaggageExt::baggage() API

The processor crate now builds successfully and is ready for crates.io publication.

+++ crates/processor/src/window/manager.rs
@@ -143,7 +143,7 @@ impl WindowMetadata {
             event_time: self.stats.max_event_time,
             processing_time: Utc::now(),
             watermark,
-            element_count: self.stats.event_count,
+            element_count: self.stats.event_count as usize,
         };
 
         let result = if let Some(wm) = watermark {
